@@ -5,8 +5,21 @@
     "dev": "vite --host",
     "build": "vite build",
     "serve": "vite preview",
-    "pbl": "scp -r dist/* root@123.56.85.24:/usr/share/nginx/html/code"
+    "pbl": "scp -r dist/* root@123.56.85.24:/usr/share/nginx/html/code",
+    "lint": "eslint --ext .js,.jsx,.ts,.tsx,.d.ts,.vue src/",
+    "lint:css": "stylelint src/**/*.{vue,scss,css,sass}",
+    "prettier": "prettier --write ."
   },
+  "lint-staged": {
+    "src/**/*.{js,jsx,ts,tsx}": "eslint",
+    "src/**/*.{vue,scss,css,sass}": "stylelint"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
+  "license": "ISC",
   "dependencies": {
     "@aligov/jssdk-mgop": "^3.1.7",
     "@ant-design/icons-vue": "^6.0.1",
@@ -17,6 +30,7 @@
     "js-cookie": "^3.0.0",
     "monaco-editor": "^0.31.1",
     "nprogress": "^0.2.0",
+    "pinia": "^2.0.14",
     "qs": "^6.10.1",
     "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-visualizer": "^5.5.0",
@@ -35,14 +49,29 @@
     "@types/qs": "^6.9.7",
     "@types/three": "^0.131.0",
     "@types/video.js": "^7.3.27",
+    "@typescript-eslint/eslint-plugin": "^5.1.0",
+    "@typescript-eslint/parser": "^5.1.0",
     "@vitejs/plugin-vue": "^1.2.2",
     "@vue/compiler-sfc": "^3.2.4",
+    "eslint": "^7.2.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.25.2",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-typescript": "^0.14.0",
+    "eslint-plugin-vue": "^7.20.0",
+    "husky": "^4.2.3",
     "naive-ui": "^2.13.0",
+    "prettier": "^2.6.2",
     "sass": "^1.37.0",
-    "typescript": "^4.1.3",
+    "stylelint": "^13.13.1",
+    "stylelint-config-standard": "^22.0.0",
+    "stylelint-scss": "^3.21.0",
+    "typescript": "^4.4.3",
     "video.js": "^7.14.3",
     "vite": "^2.5.0",
     "vite-plugin-svg-icons": "^2.0.1",
+    "vue-eslint-parser": "^7.11.0",
     "vue-tsc": "^0.1.7"
   }
 }
