@@ -1,32 +1,35 @@
-import service from '@/utils/request'
+import service from "@/utils/request";
 
-export const getTest=()=>{
+export const getTest = () => {
   return service({
-    url: '/test',
-  })
-}
-export const getTest1=()=>{
+    url: "/test",
+  });
+};
+export const getTest1 = () => {
   return service({
-    url: '/test1',
-  })
-}
-export const company=(data:any)=>{
+    url: "/test1",
+  });
+};
+export const company = (data: any) => {
   return service({
-    url: '/company',
-    method:"post",
-    data
-  })
-}
+    url: "/company",
+    method: "post",
+    data,
+  });
+};
 /**
  * binary
  * @param id
- * @returns 
+ * @returns
  */
- export const binary=(data:string,params:{team_id:number,filename:string})=>{
+export const binary = (
+  data: string,
+  params: { team_id: number; filename: string }
+) => {
   return service({
     url: "/up",
     method: "post",
     data: data,
-    params: params
+    params: params,
   });
-}
+};
