@@ -1,9 +1,9 @@
-import service from '../utils/request'
+import service from "../utils/request";
 
-export const login=(data:user)=>{
+export const login = (data: user) => {
   return service({
-    url: 'user/login',
+    url: "user/login",
     method: "post",
-    data:JSON.stringify(data)
-  })
-}
+    data: JSON.stringify(data),
+  });
+};
