@@ -5,14 +5,14 @@
 </template>
 
 <script lang="ts">
-import { defineComponent, computed } from 'vue'
+import { defineComponent, computed } from "vue";
 
 export default defineComponent({
-  name: 'SvgIcon',
+  name: "SvgIcon",
   props: {
     prefix: {
       type: String,
-      default: 'icon',
+      default: "icon",
     },
     name: {
       type: String,
@@ -20,31 +20,31 @@ export default defineComponent({
     },
     color: {
       type: String,
-      default: '#333',
+      default: "#333",
     },
     className: {
-        type: String,
-        default: ''
-      }
+      type: String,
+      default: "",
+    },
   },
   setup(props) {
-    const symbolId = computed(() => `#${props.prefix}-${props.name}`)
+    const symbolId = computed(() => `#${props.prefix}-${props.name}`);
     const svgClass = computed(() => {
-        if (props.className) {
-          return `svg-icon ${props.className}`
-        }
-        return 'svg-icon'
-      })
+      if (props.className) {
+        return `svg-icon ${props.className}`;
+      }
+      return "svg-icon";
+    });
 
-    return { symbolId,svgClass }
+    return { symbolId, svgClass };
   },
-})
+});
 </script>
 <style scope>
-  .svg-icon {
-    width: 1em;
-    height: 1em;
-    vertical-align: -0.1em; /* 因icon大小被设置为和字体大小一致，而span等标签的下边缘会和字体的基线对齐，故需设置一个往下的偏移比例，来纠正视觉上的未对齐效果 */
-    overflow: hidden;
-  }
+.svg-icon {
+  width: 1em;
+  height: 1em;
+  vertical-align: -0.1em; /* 因icon大小被设置为和字体大小一致，而span等标签的下边缘会和字体的基线对齐，故需设置一个往下的偏移比例，来纠正视觉上的未对齐效果 */
+  overflow: hidden;
+}
 </style>
