@@ -1,7 +1,9 @@
 import Layout from "@/Layout/index.vue";
-import type { RouteRecordRaw} from "vue-router"
+import type { RouteRecordRaw } from "vue-router";
+import { AddLocation } from "@element-plus/icons-vue";
+
 // 项目管理子系统-路由页面
-let router: RouteRecordRaw[] = [
+const router: RouteRecordRaw[] = [
   {
     path: "/projectManage",
     component: Layout,
@@ -13,136 +15,80 @@ let router: RouteRecordRaw[] = [
     children: [
       {
         path: "workStage",
-        component: () =>
-          import("@/views/projectManage/index.vue"),
+        component: () => import("@/views/projectManage/index.vue"),
         name: "首页",
         meta: {
-          elIcon: "el-icon-s-home",
+          icon: AddLocation,
+          title: "首页",
           roles: [],
         },
       },
       {
         path: "projectApply",
-        component: () =>
-        import("@/views/projectManage/index.vue"),
+        component: () => import("@/views/projectManage/pinia.vue"),
         props: {
           mode: "add",
         },
-        name: "项目申报",
+        name: "pinia",
         meta: {
-          elIcon: "el-icon-s-order",
+          icon: AddLocation,
           disabled: true,
+          title: "pinia",
           roles: [],
         },
       },
       {
         path: "draftBox",
-        component: () =>
-        import("@/views/projectManage/index.vue"),
+        component: () => import("@/views/projectManage/index.vue"),
         name: "草稿箱",
         meta: {
-          elIcon: "el-icon-s-cooperation",
+          icon: AddLocation,
           disabled: true,
+          title: "草稿箱",
           roles: [],
         },
       },
       {
         path: "processManage",
-        component: () =>
-        import("@/views/projectManage/index.vue"),
+        component: () => import("@/views/projectManage/index.vue"),
         name: "项目过程管理",
         meta: {
-          elIcon: "el-icon-s-cooperation",
+          icon: AddLocation,
           disabled: true,
+          title: "项目过程管理",
           roles: [],
         },
       },
       {
         path: "todo",
-        component: () =>
-        import("@/views/projectManage/index.vue"),
+        component: () => import("@/views/projectManage/index.vue"),
         name: "待办项目",
         meta: {
-          elIcon: "el-icon-s-release",
+          icon: AddLocation,
           disabled: false,
+          title: "待办项目",
           roles: [],
         },
       },
-      // {
-      //   path: "reviewManage",
-      //   name: "评审会管理",
-      //   component: () =>
-      //     import(
-      //       "@/views/application-factory/projectManage/reviewManage/index"
-      //     ),
-      //   meta: {
-      //     elIcon: "el-icon-s-grid",
-      //     disabled: false,
-      //   },
-      // },
-      // // process
-      // {
-      //   path: "process",
-      //   name: "项目过程管理",
-      //   component: () =>
-      //     import("@/views/application-factory/projectManage/process/index"),
-      //   meta: {
-      //     elIcon: "el-icon-s-grid",
-      //     disabled: false,
-      //   },
-      // },
-      // evaluate
-      // {
-      //   path: "opinion",
-      //   name: "立项意见管理",
-      //   component: () =>
-      //     import("@/views/application-factory/projectManage/opinion/index"),
-      //   meta: {
-      //     elIcon: "el-icon-s-grid",
-      //     disabled: false,
-      //   },
-      // },
-      // {
-      //   path: "evaluate",
-      //   name: "绩效评价管理",
-      //   component: () =>
-      //     import("@/views/application-factory/projectManage/evaluate/index"),
-      //   meta: {
-      //     elIcon: "el-icon-s-grid",
-      //     disabled: false,
-      //   },
-      // },
-      // maintenance
-      // {
-      //   path: "maintenance",
-      //   name: "项目运维管理",
-      //   component: () =>
-      //     import("@/views/application-factory/projectManage/maintenance/index"),
-      //   meta: {
-      //     elIcon: "el-icon-s-grid",
-      //     disabled: false,
-      //   },
-      // },
       {
         path: "alreadyDone",
-        component: () =>
-        import("@/views/projectManage/index.vue"),
+        component: () => import("@/views/projectManage/index.vue"),
         name: "已办项目",
         meta: {
-          elIcon: "el-icon-s-claim",
+          icon: AddLocation,
           disabled: false,
+          title: "已办项目",
           roles: [],
         },
       },
       {
         path: "search",
-        roles: ["xhqxmglce-sjjgly"],
-        component: () =>
-        import("@/views/projectManage/index.vue"),
+        component: () => import("@/views/projectManage/index.vue"),
         name: "项目库检索",
         meta: {
-          elIcon: "el-icon-s-grid",
+          icon: AddLocation,
           roles: [],
+          title: "项目库检索",
           // disabled: true,
         },
       },
