@@ -1,22 +1,25 @@
 // 默认缓存期限为7天
-const DEFAULT_CACHE_TIME = 60 * 60 * 24 * 7
+const DEFAULT_CACHE_TIME = 60 * 60 * 24 * 7;
 
 /**
  * 创建本地缓存对象
  * @param {string=} prefixKey -
  * @param {Object} [storage=localStorage] - sessionStorage | localStorage
  */
-export const createStorage = ({ prefixKey = '', storage = localStorage } = {}) => {
+export const createStorage = ({
+  prefixKey = "",
+  storage = localStorage,
+} = {}) => {
   /**
    * 本地缓存类
    * @class Storage
    */
   const Storage = class {
-    private storage = storage
-    private prefixKey?: string = prefixKey
+    private storage = storage;
+    private prefixKey?: string = prefixKey;
 
     private getKey(key: string) {
-      return `${this.prefixKey}${key}`.toUpperCase()
+      return `${this.prefixKey}${key}`.toUpperCase();
     }
 
     /**
@@ -28,9 +31,9 @@ export const createStorage = ({ prefixKey = '', storage = localStorage } = {}) =
     set(key: string, value: any, expire: number | null = DEFAULT_CACHE_TIME) {
       const stringData = JSON.stringify({
         value,
-        expire: expire !== null ? new Date().getTime() + expire * 1000 : null
-      })
-      this.storage.setItem(this.getKey(key), stringData)
+        expire: expire !== null ? new Date().getTime() + expire * 1000 : null,
+      });
+      this.storage.setItem(this.getKey(key), stringData);
     }
 
     /**
@@ -39,21 +42,21 @@ export const createStorage = ({ prefixKey = '', storage = localStorage } = {}) =
      * @param {*=} def 默认值
      */
     get(key: string, def: any = null) {
-      const item = this.storage.getItem(this.getKey(key))
+      const item = this.storage.getItem(this.getKey(key));
       if (item) {
         try {
-          const data = JSON.parse(item)
-          const { value, expire } = data
+          const data = JSON.parse(item);
+          const { value, expire } = data;
           // 在有效期内直接返回
           if (expire === null || expire >= Date.now()) {
-            return value
+            return value;
           }
-          this.remove(this.getKey(key))
+          this.remove(this.getKey(key));
         } catch (e) {
-          return def
+          return def;
         }
       }
-      return def
+      return def;
     }
 
     /**
@@ -61,8 +64,8 @@ export const createStorage = ({ prefixKey = '', storage = localStorage } = {}) =
      * @param {string} key
      */
     remove(key: string) {
-      console.log(key, '搜索')
-      this.storage.removeItem(this.getKey(key))
+      console.log(key, "搜索");
+      this.storage.removeItem(this.getKey(key));
     }
 
     /**
@@ -70,7 +73,7 @@ export const createStorage = ({ prefixKey = '', storage = localStorage } = {}) =
      * @memberOf Cache
      */
     clear(): void {
-      this.storage.clear()
+      this.storage.clear();
     }
 
     /**
@@ -81,8 +84,12 @@ export const createStorage = ({ prefixKey = '', storage = localStorage } = {}) =
      * 如果过期时间为设置，默认关闭浏览器自动删除
      * @example
      */
-    setCookie(name: string, value: any, expire: number | null = DEFAULT_CACHE_TIME) {
-      document.cookie = `${this.getKey(name)}=${value}; Max-Age=${expire}`
+    setCookie(
+      name: string,
+      value: any,
+      expire: number | null = DEFAULT_CACHE_TIME
+    ) {
+      document.cookie = `${this.getKey(name)}=${value}; Max-Age=${expire}`;
     }
 
     /**
@@ -90,14 +97,14 @@ export const createStorage = ({ prefixKey = '', storage = localStorage } = {}) =
      * @param name
      */
     getCookie(name: string): string {
-      const cookieArr = document.cookie.split('; ')
+      const cookieArr = document.cookie.split("; ");
       for (let i = 0, length = cookieArr.length; i < length; i++) {
-        const kv = cookieArr[i].split('=')
+        const kv = cookieArr[i].split("=");
         if (kv[0] === this.getKey(name)) {
-          return kv[1]
+          return kv[1];
         }
       }
-      return ''
+      return "";
     }
 
     /**
@@ -105,24 +112,24 @@ export const createStorage = ({ prefixKey = '', storage = localStorage } = {}) =
      * @param {string} key
      */
     removeCookie(key: string) {
-      this.setCookie(key, 1, -1)
+      this.setCookie(key, 1, -1);
     }
 
     /**
      * 清空cookie，使所有cookie失效
      */
     clearCookie(): void {
-      const keys = document.cookie.match(/[^ =;]+(?==)/g)
+      const keys = document.cookie.match(/[^ =;]+(?==)/g);
       if (keys) {
         for (let i = keys.length; i--; ) {
-          document.cookie = keys[i] + '=0;expire=' + new Date(0).toUTCString()
+          document.cookie = keys[i] + "=0;expire=" + new Date(0).toUTCString();
         }
       }
     }
-  }
-  return new Storage()
-}
+  };
+  return new Storage();
+};
 
-export const Storage = createStorage()
+export const Storage = createStorage();
 
-export default Storage
+export default Storage;
