@@ -1,15 +1,14 @@
 <template>
-    <div class="Drag" >
-            <slot></slot>
-    </div>
+  <div class="Drag">
+    <slot></slot>
+  </div>
 </template>
-<script lang="ts" setup>
-</script>
+<script lang="ts" setup></script>
 <style lang="scss">
-.Drag{
-    width: 100%;
-    height: 100%;
-    position: relative;
-    // background:pink;
+.Drag {
+  width: 100%;
+  height: 100%;
+  position: relative;
+  // background:pink;
 }
 </style>
