@@ -1,23 +1,23 @@
 <template>
-	<div :class="classes">
-		<slot></slot>
-	</div>
+  <div :class="classes">
+    <slot></slot>
+  </div>
 </template>
 
 <script lang="ts">
-  export default {
-    name: 'Pane',
-    props: {
-      className: String
-    },
-    data() {
-      const classes = [this.$parent.split, this.className]
-      return {
-        classes: classes.join(' '),
-        percent: 50
-      }
-    }
-  }
+export default {
+  name: "Pane",
+  props: {
+    className: String,
+  },
+  data() {
+    const classes = [this.$parent.split, this.className];
+    return {
+      classes: classes.join(" "),
+      percent: 50,
+    };
+  },
+};
 </script>
 
 <style scoped>
