@@ -1,26 +1,27 @@
 <template>
-    <div>子组件b
-      <ChuildC />
-    </div>
+  <div>
+    子组件b
+    <ChuildC />
+  </div>
 </template>
-<script lang="ts" >
+<script lang="ts">
 import { defineComponent } from "vue";
 import ChuildC from "./chuildC.vue";
 export default defineComponent({
- components:{
-    ChuildC
- },
-  created(){
-    console.log("我是B组件我执行了,created")
+  components: {
+    ChuildC,
   },
-  mounted(){
-    console.log("我是B组件我执行了,mounted")
+  created() {
+    console.log("我是B组件我执行了,created");
   },
-  beforeDestory(){
-    console.log("我是B组件我执行了,beforeDestory")
+  mounted() {
+    console.log("我是B组件我执行了,mounted");
   },
-  destoryed(){
-        console.log("我是B组件我执行了,destoryed")
-  }
-})
+  beforeDestory() {
+    console.log("我是B组件我执行了,beforeDestory");
+  },
+  destoryed() {
+    console.log("我是B组件我执行了,destoryed");
+  },
+});
 </script>
